requirements.txt
streamlit
pandas
google-generativeai>=0.8.3
gspread
oauth2client
Pillow
altair
